Move website setup to shell script.
This does the exact same things as the old chebsite_setup.m, plus it
checks for a few other potential issues, namely, a missing Git, Python,
or pip and an out-of-date python_dateutil module.

I rewrote this in shell because coding all of this in MATLAB using
system() and the like was getting annoying.  (Really, this should
probably have been in shell in the first place.)  As an added benefit,
we don't have to worry about the "git clone" commands hooking into
MATLAB's libcurl instead of the system one anymore.